rename "categoryList_characteristics" to "CATEGORY"
1. import changed in all files where this object is used
