+++ controllers/card.js
@@ -1,7 +1,7 @@
 const Product = require("../models/Product");
 
 const { CATALOG } = require("../db/catalog");
-const catalog_option = require("../db/catalog_characteristics");
+const { CATEGORY } = require("../db/category");
 
 module.exports.getByIdCard = async function (req, res) {
   console.log("Server getByIdCard");
@@ -109,14 +109,12 @@ module.exports.getByIdCardCharacteristics = async function (req, res) {
 
       if (product.category.length === 3) {
         characteristicsName =
-          catalog_option.categoryList_characteristics[product.category[0]]
-            .nameListCategory[product.category[1]].subNameListCategory[
-            product.category[2]
-          ].characteristics;
+          CATEGORY[product.category[0]].nameListCategory[product.category[1]]
+            .subNameListCategory[product.category[2]].characteristics;
       } else if (product.category.length === 2) {
         characteristicsName =
-          catalog_option.categoryList_characteristics[product.category[0]]
-            .nameListCategory[product.category[1]].characteristics;
+          CATEGORY[product.category[0]].nameListCategory[product.category[1]]
+            .characteristics;
       }
 
       const productCharacteristics = Object.assign(
